Stack_seal_mitigation: Seal stack for interrupt deprivileging
Seal the handler mode stack (MSP) in case of interrupt deprivileging
for Function mode. The interrupt deprivileging is acheived by the ISR
calling an SVC and a seal is applied on MSP before the interrupt is
deprivileged. Other exceptions such as faults runs to panic so there
is no need of sealing caller contexts for these exceptions.

This patch also removes the separate inline assembly implementation for
Mainline as the benefit of a single code to maintain and patch far
outweigh any benefits gained by a separate implementation.

Change-Id: I8fdff140f4157882853e3abd24cc23b7e68ee0a3
Signed-off-by: Ken Liu <ken.liu@arm.com>
Signed-off-by: Mate Toth-Pal <mate.toth-pal@arm.com>
Signed-off-by: Soby Mathew <soby.mathew@arm.com>